requirements.txt
pygame
paho-mqtt